third_party/blink/web_tests/fast/css/outline-currentcolor-expected.txt

Test that outline-color is non-inherit and currentColor is handled correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(test1).outlineColor is "rgb(0, 128, 0)"
PASS getComputedStyle(test2).outlineColor is "rgb(0, 128, 0)"
PASS getComputedStyle(test3).outlineColor is "rgb(0, 128, 0)"
PASS getComputedStyle(test4).outlineColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
